Support Dissimilarity Index
Previously I was unable to find examples of diffs that had dissimilarity
index lines on them. I was able to reproduce this with a specific git
command (command mentioned in a previous commit).

I'm not 100% confident about the DAG configuration for this node
regarding it's possible parent and children nodes. However, it works for
the examples I've seen.