Fix publishing to maven central
The old method of publishing no longer worked due to a number of backend
changes. There are a lot of details, but this is a good starting point:

https://central.sonatype.org/pages/ossrh-eol/

In order to publish we needed to start using jreleaser which also
involved upgrading gradle. We also are automatically publishing now
which was a manual process before.